requirements/dev.txt
-r test.txt

django-debug-toolbar==3.2.1
ipython==7.25.0
djlint

